backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
arq>=0.26.1
PyJWT>=2.8.0
bcrypt>=4.0.0
email-validator>=2.1.0
pgvector>=0.3.6
openai>=1.30.0
anthropic>=0.28.0
structlog>=24.1.0
greenlet>=3.0.3
